flight_objects.py is the script that actually sets up a FlightDatabase w sample things
